Keep key attributes along with full alerts object
Not a huge change; really just adds back the `alerts` attribute. I included the `alerts_sting` json only as an easy reference for how to access the information in the `alerts` attribute object, but the `alerts` attribute is the part that's actually useful for templates, automations, etc.

Other than that the only real change is that I used the ternary operator for assigning the attributes, just to simplify the code a bit. I made sure to keep all of the original attributes in place so that, as you mentioned, nothing would break for the users and they could more easily access the most important alert.